Warnings from Linnea
git-svn-id: svn+ssh://svn.gecode.org/srv/gecode/svn/gecode/trunk@16001 64335634-5103-0410-b293-fc3d331e086d

+++ test/int/bin-packing.cpp
@@ -299,7 +299,7 @@ namespace Test { namespace Int {
           delete home;
           return false;
         }
-        if (clique.size() != mc.size()) {
+        if (static_cast<unsigned int>(clique.size()) != mc.size()) {
           delete home;
           return false;
         }
